feat: move user query to a separate file
