+++ src/gql/query/index.ts
@@ -20,5 +20,6 @@ export * from "./spotPools"
 export * from "./spotPoolSwap"
 export * from "./staking"
 export * from "./stats"
+export * from "./user"
 export * from "./users"
 export * from "./wasm"
