+++ src/gql/query/users.ts
@@ -7,27 +7,10 @@ import {
   GQLUser,
   GQLUserOrder,
   DeepPartial,
-  GQLQueryGqlUserArgs,
 } from ".."
 
-export interface GqlOutUser {
-  user?: GQLQuery["user"]
-}
 export interface GqlOutUsers {
   users?: GQLQuery["users"]
-}
-
-export const userQueryString = (
-  args: GQLQueryGqlUserArgs,
-  excludeParentObject: boolean,
-  fields: DeepPartial<GQLUser>
-) => {
-  return gqlQuery(
-    "user",
-    args,
-    convertObjectToPropertiesString(fields),
-    excludeParentObject
-  )
 }
 
 export const usersQueryString = (
@@ -46,13 +29,6 @@ export const usersQueryString = (
     excludeParentObject
   )
 }
-
-export const user = async (
-  args: GQLQueryGqlUserArgs,
-  endpt: string,
-  fields: DeepPartial<GQLUser>
-): Promise<GqlOutUser> =>
-  doGqlQuery(userQueryString(args, false, fields), endpt)
 
 export const users = async (
   args: GQLQueryGqlUsersArgs,
